Update
Updated to Current Version.

+++ Scripts/Items/Addons/Craft Addons/CraftAddonDeed.cs
@@ -1,3 +1,11 @@
+using System;
+using Server;
+using Server.Engines.Craft;
+using System.Collections.Generic;
+using Server.Multis;
+using Server.Mobiles;
+using System.Linq;
+
 namespace Server.Items
 {
     public abstract class CraftAddonDeed : BaseAddonDeed
@@ -26,7 +34,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0);
+            writer.Write((int)0);
 
             writer.Write(UsesRemaining);
         }
