+++ Scripts/Items/Artifacts/Equipment/Clothing/HoodedShroudOfShadows.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace Server.Items
 {
     [Flipable(0x2684, 0x2683)]
     public class HoodedShroudOfShadows : BaseOuterTorso
-    {
+	{
         [Constructable]
         public HoodedShroudOfShadows()
             : this(0x455)
@@ -32,7 +34,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
