+++ Scripts/Items/Books/TanBook.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Server.Items
 {
     public class TanBook : BaseBook
@@ -42,7 +44,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
     }
 }
