+++ Scripts/Items/Consumables/Kindling.cs
@@ -1,6 +1,7 @@
+using System;
+using System.Collections;
 using Server.Network;
 using Server.Regions;
-using System.Collections;
 
 namespace Server.Items
 {
@@ -30,7 +31,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(1); // version
+            writer.Write((int)1); // version
         }
 
         public override void Deserialize(GenericReader reader)
