+++ Scripts/Items/Decorative/Bridle.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Server.Items
 {
     [TypeAlias("Server.Items.DecoBridle")]
@@ -19,7 +21,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(0);
+            writer.Write((int)0);
         }
 
         public override void Deserialize(GenericReader reader)
@@ -48,7 +50,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(0);
+            writer.Write((int)0);
         }
 
         public override void Deserialize(GenericReader reader)
