+++ Scripts/Items/Decorative/DarkWoodWall.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Server.Items
 {
     public enum DarkWoodWallTypes
@@ -41,7 +43,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
