+++ Scripts/Items/Decorative/ExecutionersCap.cs
@@ -1,19 +1,21 @@
+using System;
+
 namespace Server.Items
 {
     public class ExecutionersCap : Item
     {
-        [Constructable]
+		[Constructable]
         public ExecutionersCap()
             : this(1)
         {
         }
-
+		
         [Constructable]
         public ExecutionersCap(int amount)
             : base(0xF83)
         {
-            Stackable = true;
-            Amount = amount;
+			Stackable = true;
+			Amount = amount;
             Weight = 1.0;
         }
 
@@ -25,7 +27,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(0);
+            writer.Write((int)0);
         }
 
         public override void Deserialize(GenericReader reader)
