+++ Scripts/Items/Decorative/Fur.cs
@@ -1,4 +1,5 @@
 using Server.Mobiles;
+using System;
 
 namespace Server.Items
 {
@@ -37,7 +38,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
