+++ Scripts/Items/Decorative/Sign.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Server.Items
 {
     public enum SignFacing
@@ -100,7 +102,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0);
+            writer.Write((int)0);
         }
 
         public override void Deserialize(GenericReader reader)
