+++ Scripts/Items/Decorative/SpikeHead.cs
@@ -1,3 +1,4 @@
+using System;
 using Server.Gumps;
 
 namespace Server.Items
@@ -59,9 +60,9 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(0); // version
+            writer.Write((int)0); // version
 
-            writer.Write(GumpID);
+            writer.Write((int)GumpID);
         }
 
         public override void Deserialize(GenericReader reader)
