+++ Scripts/Items/Decorative/ThinStoneWall.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Server.Items
 {
     public enum ThinStoneWallTypes
@@ -46,7 +48,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
