+++ Scripts/Items/Decorative/WhiteStoneWall.cs
@@ -6,6 +6,8 @@
 * CREATED : 10-07.2002                 *
 * **************************************/
 
+using System;
+
 namespace Server.Items
 {
     public enum WhiteStoneWallTypes
@@ -64,7 +66,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
