+++ Scripts/Items/Functional/SeedBox/SeedEntry.cs
@@ -1,3 +1,7 @@
+using System;
+using Server;
+using Server.Engines.Plants;
+
 namespace Server.Engines.Plants
 {
     public class SeedEntry
@@ -24,7 +28,7 @@ namespace Server.Engines.Plants
 
         public void Serialize(GenericWriter writer)
         {
-            writer.Write(0);
+            writer.Write((int)0);
 
             writer.Write(Seed);
             writer.Write(Image);
