+++ Scripts/Items/Functional/SlidingDoors.cs
@@ -1,5 +1,7 @@
+using System;
+
 namespace Server.Items
-{
+{ 
     public class PaperSlidingDoor : BaseDoor
     {
         [Constructable]
@@ -17,7 +19,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
@@ -45,7 +47,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
@@ -73,7 +75,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
