+++ Scripts/Items/Resource/Blackrock.cs
@@ -1,3 +1,6 @@
+using System;
+using Server.Network;
+
 namespace Server.Items
 {
     public class Blackrock : Item
@@ -26,12 +29,12 @@ namespace Server.Items
                 }
             }
         }
-
+		
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
