+++ Scripts/Items/StoreBought/GreenGoblinStatuette.cs
@@ -1,3 +1,6 @@
+using System;
+using Server.Gumps;
+
 namespace Server.Items
 {
     public class GreenGoblinStatuette : MonsterStatuette
@@ -118,7 +121,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write(1);
+            writer.Write((int)1);
             writer.Write(User);
         }
 
@@ -128,14 +131,14 @@ namespace Server.Items
 
             int version = reader.ReadInt();
 
-            switch (version)
+            switch(version)
             {
                 case 1:
                     User = reader.ReadMobile();
                     break;
             }
 
-            if (User != null)
+            if(User != null)
             {
                 TurnOn(User);
             }
