+++ Scripts/Mobiles/Normal/EnslavedSatyr.cs
@@ -1,10 +1,11 @@
+using System;
 using Server.Items;
 
 namespace Server.Mobiles
 {
-    [CorpseName("an enslaved satyr corpse")]
+    [CorpseName("an enslaved satyr corpse")] 
     public class EnslavedSatyr : Satyr
-    {
+    { 
         [Constructable]
         public EnslavedSatyr()
             : base()
@@ -24,23 +25,23 @@ namespace Server.Mobiles
 
         public override void OnDeath(Container c)
         {
-            base.OnDeath(c);
-
-            if (Utility.RandomDouble() < 0.1)
-                c.DropItem(new ParrotItem());
+            base.OnDeath(c);		
+			
+            if (Utility.RandomDouble() < 0.1)				
+                c.DropItem(new ParrotItem());	
         }
 
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-
-            writer.Write(0); // version
+			
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
         {
             base.Deserialize(reader);
-
+			
             int version = reader.ReadInt();
         }
     }
