+++ Scripts/Mobiles/Normal/FrostDrake.cs
@@ -1,3 +1,7 @@
+using System;
+using Server;
+using Server.Items;
+
 namespace Server.Mobiles
 {
     [CorpseName("a drake corpse")]
@@ -17,7 +21,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(0);
+            writer.Write((int)0);
         }
 
         public override void Deserialize(GenericReader reader)
