+++ Scripts/Mobiles/Normal/SpeckledScorpion.cs
@@ -1,3 +1,4 @@
+using System;
 using Server.Items;
 
 namespace Server.Mobiles
@@ -20,8 +21,8 @@ namespace Server.Mobiles
 
         public override void OnDeath(Container c)
         {
-            base.OnDeath(c);
-
+            base.OnDeath(c);	
+			
             if (Utility.RandomDouble() < 0.4)
                 c.DropItem(new SpeckledPoisonSac());
         }
@@ -29,14 +30,14 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-
-            writer.Write(0); // version
+			
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
         {
             base.Deserialize(reader);
-
+			
             int version = reader.ReadInt();
         }
     }
