+++ Scripts/Quests/CloakOfHumility/Items/GoldShield.cs
@@ -1,3 +1,9 @@
+using System;
+using Server.Engines.Quests;
+using Server.Mobiles;
+using System.Collections.Generic;
+using Server.Gumps;
+
 namespace Server.Items
 {
     public class GoldShield : OrderShield
@@ -13,12 +19,12 @@ namespace Server.Items
             : base(serial)
         {
         }
-
+ 
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
 
-            writer.Write(0); // version
+            writer.Write((int)0); // version
         }
 
         public override void Deserialize(GenericReader reader)
