+++ Scripts/Services/Seasonal Events/KrampusEncounter/Items/KrampusPunishinList.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Server.Items
 {
     [Flipable(0x46AE, 0x46AF)]
@@ -26,7 +28,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(0); // version
+            writer.Write((int)0); // version
 
             writer.Write(_Name);
         }
